feat:实现了浏览量统计，加入了用户身份验证

+++ my_blog/article/views.py
@@ -33,6 +33,10 @@ def article_detail(request, id):
     # 取出相应的文章
     article = ArticlePost.objects.get(id=id)
 
+    # 浏览量+1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
+
     # 将markdown语法渲染成html样式
     article.body = markdown.markdown(article.body,
                                      extensions=[
@@ -79,9 +83,15 @@ def article_create(request):
 
 
 # 删除文章
+# 检查登录
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+
+    # 过滤非作者用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权删除这篇文章")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成后返回文章列表
@@ -97,7 +107,9 @@ def article_safe_delete(request,id):
     else:
         return HttpResponse("仅允许post请求")
 
-
+# 检查登录
+# 更新文章
+@login_required(login_url='/userprofile/login/')
 def article_update(request,id):
     """
     更新文章的视图函数
@@ -108,6 +120,10 @@ def article_update(request,id):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+
+    # 过滤非作者用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章")
     # 判断用户是否为POST提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
